+++ digicampipe/reco/camera/r0.py
